requirements.txt
# --- Procesamiento de Documentos y Datos ---
pandas==2.2.2
pdfplumber==0.11.1
numpy==1.26.4
PyYAML==6.0.1
tabulate==0.9.0

# --- LangChain y OpenAI (VERSIONES CON Pydantic FIJO) ---
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
langchain-openai==0.1.6
openai==1.24.1
pydantic==2.7.4  # <-- LA CLAVE: Forzamos la versión 2

# --- Base de Datos Vectorial ---
faiss-cpu==1.8.0

# --- Visualización de Datos ---
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.22.0

# --- Utilidades ---
python-dotenv==1.0.1

# --- Machine Learning (Usado en validation.py) ---
scikit-learn==1.5.0